Update git search to include multi_dot

+++ SearchGit/ExtractedFunctions/ultralytics_ultralytics/ultralytics-main_ultralytics_nn_modules_transformer.py_build_2d_sincos_position_embedding.py
@@ -20,7 +20,8 @@
         omega = torch.arange(pos_dim, dtype=torch.float32) / pos_dim
         omega = 1.0 / (temperature**omega)
 
-        out_w = grid_w.flatten()[..., None] @ omega[None]
-        out_h = grid_h.flatten()[..., None] @ omega[None]
+        # Pin matmul to fp32 for CoreML export: fp16 sin/cos on integer-derived positions accumulates visible error.
+        out_w = grid_w.flatten()[..., None].float() @ omega[None]
+        out_h = grid_h.flatten()[..., None].float() @ omega[None]
 
         return torch.cat([torch.sin(out_w), torch.cos(out_w), torch.sin(out_h), torch.cos(out_h)], 1)[None]
